chore: move social-network to social-network-provider
