+++ README.md
@@ -62,3 +62,4 @@ For general introductions, see https://Maskbook.com/
 -   ./protocols - Some standard definitions of protocols that used in Maskbook
 -   ./network - How does Maskbook exchange information between users
 -   ./social-network/ - Maskbook's multiple network support (Work in progress)
+-   ./social-network-provider/ - Maskbook's multiple network support (Work in progress)
