+++ src/components/InjectedComponents/DecryptedPost.tsx
@@ -10,6 +10,9 @@ import { Link, Box, useMediaQuery, useTheme } from '@material-ui/core'
 import { Person } from '../../database'
 import { PersonIdentifier } from '../../database/type'
 import { NotSetupYetPrompt } from './NotSetupYetPrompt'
+import { myUsernameRef } from '../../extension/content-script/injections/MyUsername'
+import { isMobile } from '../../social-network-provider/facebook.com/isMobile'
+import { useValueRef } from '../../utils/hooks/useValueRef'
 
 interface DecryptPostSuccessProps {
     data: { signatureVerifyResult: boolean; content: string }
