+++ src/extension/content-script/injections/Posts.tsx
@@ -4,7 +4,7 @@ import { getPersonIdentifierAtFacebook } from './MyUsername'
 import { renderInShadowRoot } from '../../../utils/jss/renderInShadowRoot'
 import { deconstructPayload } from '../../../utils/type-transform/Payload'
 import { PersonIdentifier } from '../../../database/type'
-import { isMobile } from '../../../social-network/facebook.com/isMobile'
+import { isMobile } from '../../../social-network-provider/facebook.com/isMobile'
 import { PostCommentDecrypted, PostComment } from '../../../components/InjectedComponents/PostComments'
 import { useValueRef } from '../../../utils/hooks/useValueRef'
 import { PostInspector } from './Posts/PostInspector'
