+++ src/social-network-provider/facebook.com/isMobile.ts
